+++ packages/columns/src/lib/columns.spec.ts
@@ -1,0 +1,7 @@
+import { ColumnsExtension } from './columns';
+
+describe('columns', () => {
+  it('should work', () => {
+    expect(ColumnsExtension.name).toEqual('columnsExtension');
+  });
+});