+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## v2.1.1
+
+
+### 🚀 Features
+
+- add columns extension
+
+### ❤️  Thank You
+
+- Layton Whiteley
+
 ## v2.1.0
 
 ## v2.0.0